Don't trigger walk error on non-existing directories in dpkg analyzer
This change reverts parts of the the changes done in commit:

Re-use existing `ErrorCallback` WalkOption in post-analysis fs walks

Reading the content of the /var/lib/dpkg/info directory is
optional and the /var/lib/dpkg/status.d directory is particularly
uncommon. We should not trigger errors when these directory are
not existing.