tar: avoid unbound list of skipped dirs
The way skipped dirs are implemented when going through a tar file
was in O(found skipped dirs). In the case of tar files with lots
of skipped directories, the underSkippedDirs could take up to 14%
of our CPU profiles.

Instead of accumulating skipped directories, let's check for every
file if any of its parent directories is skipped.

This way we do not have an unbounded loop, and we also avoid many
of the allocations done by path cleaning (using filepath.Rel).